Kaggle/requirements.txt
numpy==1.26.4
pandas==2.2.1
pillow==10.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
datasets 
transformers 
accelerate
torch
torchvision
scikit-learn